requirements.txt
Babel==2.7.0
Django==2.2.5
django-phonenumber-field==3.0.1
mysqlclient==1.4.4
phonenumbers==8.10.19
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
